Customize Delimitation Colors

delim_brewer() returns a set of colors created by interpolating or using color palettes from RColorBrewer, viridisLite or randomcoloR.

Usage

delim_brewer(delim, package = NULL, palette = NULL, seed = NULL)

Arguments

delim

Output from delim_join.

package

Package which contains color palettes. Available options are "RColorBrewer", "viridisLite" or "randomcoloR".

palette

A palette name. brewer.pal for RColorBrewer or viridis for viridisLite options.

seed

Integer. Number to initialize random number generator.

Value

A character vector of hexadecimal color codes.

Details

delim_brewer() interpolates over a color palette and returns a vector of random colors whose length is equal to the sum of unique species delimitation partitions in delim. For reproducibility, make sure to provide a seed. If not provided, Sys.time will be used as seed instead. One should also try different seeds to get best color combinations for plotting.

Author

Rupert A. Collins, Pedro S. Bittencourt

Examples

# create a vector of colors
cols <- delim_brewer(geophagus_delims, package = "randomcoloR")
#> Warning: ⚠ Argument `seed` not found. Using `Sys.time()` as seed.
#> ℹ For reproducibility, you may want to set a custom `seed` instead. `seed` is
#>   printed below:
#> 2025-04-10 18:52:40.674542