Turns bGMYC Results Into a Tibble

bgmyc_tbl() processes output from bgmyc.singlephy into an object of class tbl_df.

Usage

bgmyc_tbl(bgmyc_res, ppcutoff = 0.05, delimname = "bgmyc")

Source

Reid N.M., Carstens B.C. 2012. Phylogenetic estimation error can decrease the accuracy of species delimitation: a Bayesian implementation of the general mixed Yule-coalescent model. BMC Evolutionary Biology 12 (196).

Arguments

bgmyc_res

Output from bgmyc.singlephy.

ppcutoff

Posterior probability threshold for clustering samples into species partitions. See bgmyc.point for details. Default to 0.05.

delimname

Character. String to rename the delimitation method in the table. Default to 'bgmyc'.

Value

an object of class tbl_df.

Details

bGMYC package uses spec.probmat to create a matrix of probability of conspecificity and bgmyc.point to split samples into a list which individuals meets the threshold specified by ppcutoff. bgmyc_tbl() wraps up these two functions into a single one and turns these inputs into a tibble which matches the output from gmyc_tbl and locmin_tbl.

Author

Noah M. Reid.

Examples

if (FALSE) { # \dontrun{

# run bGMYC
bgmyc_res <- bGMYC::bgmyc.singlephy(geophagus_beast,
  mcmc = 11000,
  burnin = 1000,
  thinning = 100,
  t1 = 2,
  t2 = ape::Ntip(geophagus_beast),
  start = c(1, 0.5, 50)
)

# create a tibble
bgmyc_df <- bgmyc_tbl(bgmyc_res, ppcutoff = 0.05)

# check
bgmyc_df
} # }