Turns GMYC Results Into a Tibble

gmyc_tbl() processes output from gmyc into an object of class tbl_df.

Usage

gmyc_tbl(gmyc_res, delimname = "gmyc")

Source

Pons J., Barraclough T. G., Gomez-Zurita J., Cardoso A., Duran D. P., Hazell S., Kamoun S., Sumlin W. D., Vogler A. P. 2006. Sequence-based species delimitation for the DNA taxonomy of undescribed insects. Systematic Biology. 55:595-609.

Monaghan M. T., Wild R., Elliot M., Fujisawa T., Balke M., Inward D. J. G., Lees D. C., Ranaivosolo R., Eggleton P., Barraclough T. G., Vogler A. P. 2009. Accelerated species inventory on Madagascar using coalescent-based models of species delineation. Systematic Biology. 58:298-311.

Fujisawa T., Barraclough T. G. 2013. Delimiting Species Using Single-Locus Data and the Generalized Mixed Yule Coalescent Approach: A Revised Method and Evaluation on Simulated Data Sets. Systematic Biology. 62(5):707–724.

Arguments

gmyc_res

Output from gmyc.

delimname

Character. String to rename the delimitation method in the table. Default to 'gmyc'.

Value

An object of class tbl_df.

Details

splits package uses gmyc to optimize genetic clusters and spec.list to cluster samples into species partitions. gmyc_tbl() turns these results into a tibble which matches the output from bgmyc_tbl and locmin_tbl.

Author

Thomas Ezard, Tomochika Fujisawa, Tim Barraclough.

Examples

if (FALSE) { # \dontrun{

# run GMYC
gmyc_res <- splits::gmyc(ape::as.phylo(geophagus_beast))

# create a tibble
gmyc_df <- gmyc_tbl(gmyc_res)

# check
gmyc_df
} # }